rename: multiview demo
